feat: full test coverage on macos & ios, v8 14 (#28)
- Adds full test coverage 100% success for iOS, and introduces macOS test support too
- Introduces some built-in Node.js APIs like node:path, etc. A process global is also added with minimal API surface
- Bunch of fixes in the FFI layer, related to correctness & performance
- Introduces Generated Signature Dispatch: compile time generated trampolines that boost performance of native calls by 20% on average through eliminating libffi layer
- Improved memory management, less leaks. Introduces tests for reliable memory management
- Removes types, build artifacts and metadata binaries from the repo
- Cleans up old runtime related stuff
- Fetches prebuilt V8 binaries now. To be added for Hermes soon.

+++ NativeScript/napi/v8/jsr.h
@@ -6,6 +6,7 @@
 #define TEST_APP_JSR_H
 
 #include <memory>
+
 #include "SimpleAllocator.h"
 #include "jsr_common.h"
 #include "libplatform/libplatform.h"
@@ -97,7 +98,7 @@ namespace tns {
 inline static v8::Local<v8::String> ConvertToV8String(v8::Isolate* isolate,
                                                       const std::string& s) {
   return v8::String::NewFromUtf8(isolate, s.c_str(), v8::NewStringType::kNormal,
-                                 s.length())
+                                 static_cast<int>(s.length()))
       .ToLocalChecked();
 }
